Combine single-node and multi-node QA modules
