+++ x-pack/plugin/esql/qa/base/src/main/java/org/elasticsearch/xpack/esql/qa/rest/generative/GenerativeRestTest.java
