+++ x-pack/plugin/esql/qa/heap-attack/build.gradle
@@ -1,0 +1,9 @@
+apply plugin: 'elasticsearch.internal-java-rest-test'
+
+dependencies {
+  javaRestTestImplementation project(xpackModule('esql:qa:base'))
+}
+
+tasks.named('javaRestTest') {
+  usesDefaultDistribution()
+}
