--- x-pack/plugin/esql/qa/server/mixed-cluster/build.gradle
@@ -1,35 +1,0 @@
-
-import org.elasticsearch.gradle.Version
-import org.elasticsearch.gradle.VersionProperties
-import org.elasticsearch.gradle.internal.info.BuildParams
-import org.elasticsearch.gradle.testclusters.StandaloneRestIntegTestTask
-
-apply plugin: 'elasticsearch.internal-java-rest-test'
-apply plugin: 'elasticsearch.bwc-test'
-
-restResources {
-  restApi {
-    include '_common', 'bulk', 'get', 'indices', 'esql', 'xpack', 'enrich', 'cluster'
-  }
-  restTests {
-    includeXpack 'esql'
-  }
-}
-
-dependencies {
-  javaRestTestImplementation project(xpackModule('esql:qa:testFixtures'))
-  javaRestTestImplementation project(xpackModule('esql:qa:server'))
-}
-
-def supportedVersion = bwcVersion -> {
-  // ESQL is available in 8.11 or later
-  return bwcVersion.onOrAfter(Version.fromString("8.11.0"));
-}
-
-BuildParams.bwcVersions.withWireCompatible(supportedVersion) { bwcVersion, baseName ->
-  tasks.register(bwcTaskName(bwcVersion), StandaloneRestIntegTestTask) {
-    usesBwcDistribution(bwcVersion)
-    systemProperty("tests.old_cluster_version", bwcVersion)
-    maxParallelForks = 1
-  }
-}
