--- x-pack/plugin/esql/qa/server/multi-node/build.gradle
@@ -1,11 +1,0 @@
-apply plugin: 'elasticsearch.internal-java-rest-test'
-
-dependencies {
-  javaRestTestImplementation project(xpackModule('esql:qa:testFixtures'))
-  javaRestTestImplementation project(xpackModule('esql:qa:server'))
-}
-
-
-tasks.named('javaRestTest') {
-  usesDefaultDistribution()
-}
