--- x-pack/plugin/esql/qa/server/single-node/build.gradle
@@ -1,27 +1,0 @@
-apply plugin: 'elasticsearch.internal-java-rest-test'
-apply plugin: 'elasticsearch.internal-yaml-rest-test'
-
-dependencies {
-  javaRestTestImplementation project(xpackModule('esql:qa:testFixtures'))
-  javaRestTestImplementation project(xpackModule('esql:qa:server'))
-  yamlRestTestImplementation project(xpackModule('esql:qa:server'))
-}
-
-restResources {
-  restApi {
-    include '_common', 'bulk', 'get', 'indices', 'esql', 'xpack', 'enrich', 'cluster'
-  }
-  restTests {
-    includeXpack 'esql'
-  }
-}
-
-tasks.named('javaRestTest') {
-  usesDefaultDistribution()
-  maxParallelForks = 1
-}
-
-tasks.named('yamlRestTest') {
-  usesDefaultDistribution()
-  maxParallelForks = 1
-}
