+++ x-pack/plugin/esql/qa/single-cluster/build.gradle
@@ -1,0 +1,27 @@
+apply plugin: 'elasticsearch.internal-java-rest-test'
+apply plugin: 'elasticsearch.internal-yaml-rest-test'
+
+dependencies {
+  javaRestTestImplementation project(xpackModule('esql:qa:testFixtures'))
+  javaRestTestImplementation project(xpackModule('esql:qa:base'))
+  yamlRestTestImplementation project(xpackModule('esql:qa:base'))
+}
+
+restResources {
+  restApi {
+    include '_common', 'bulk', 'get', 'indices', 'esql', 'xpack', 'enrich', 'cluster'
+  }
+  restTests {
+    includeXpack 'esql'
+  }
+}
+
+tasks.named('javaRestTest') {
+  usesDefaultDistribution()
+  maxParallelForks = 1
+}
+
+tasks.named('yamlRestTest') {
+  usesDefaultDistribution()
+  maxParallelForks = 1
+}
